fork old dev overlay and render conditionally
