+++ packages/next/src/client/components/react-dev-overlay/app/ReactDevOverlay.tsx
@@ -1,92 +1,18 @@
-import React from 'react'
-import type { OverlayState } from '../shared'
-import { ShadowPortal } from '../internal/components/ShadowPortal'
-import { BuildError } from '../internal/container/BuildError'
-import { Errors } from '../internal/container/Errors'
-import { StaticIndicator } from '../internal/container/StaticIndicator'
-import { Base } from '../internal/styles/Base'
-import { ComponentStyles } from '../internal/styles/ComponentStyles'
-import { CssReset } from '../internal/styles/CssReset'
-import { RootLayoutMissingTagsError } from '../internal/container/root-layout-missing-tags-error'
 import type { Dispatcher } from './hot-reloader-client'
-import { RuntimeErrorHandler } from '../internal/helpers/runtime-error-handler'
+import type { OverlayState } from '../shared'
+import OldDevOverlay from './OldDevOverlay'
+import NewDevOverlay from '../_new/app/NewDevOverlay'
 
-interface ReactDevOverlayState {
-  isReactError: boolean
+type ReactDevOverlayProps = {
+  state: OverlayState
+  dispatcher?: Dispatcher
+  children: React.ReactNode
 }
-export default class ReactDevOverlay extends React.PureComponent<
-  {
-    state: OverlayState
-    dispatcher?: Dispatcher
-    children: React.ReactNode
-  },
-  ReactDevOverlayState
-> {
-  state = { isReactError: false }
 
-  static getDerivedStateFromError(error: Error): ReactDevOverlayState {
-    if (!error.stack) return { isReactError: false }
+export default function ReactDevOverlay(props: ReactDevOverlayProps) {
+  const Overlay = !!process.env.__NEXT_EXPERIMENTAL_NEW_DEV_OVERLAY
+    ? NewDevOverlay
+    : OldDevOverlay
 
-    RuntimeErrorHandler.hadRuntimeError = true
-    return {
-      isReactError: true,
-    }
-  }
-
-  render() {
-    const { state, children, dispatcher } = this.props
-    const { isReactError } = this.state
-
-    const hasBuildError = state.buildError != null
-    const hasRuntimeErrors = Boolean(state.errors.length)
-    const hasStaticIndicator = state.staticIndicator
-    const debugInfo = state.debugInfo
-
-    return (
-      <>
-        {isReactError ? (
-          <html>
-            <head></head>
-            <body></body>
-          </html>
-        ) : (
-          children
-        )}
-        <ShadowPortal>
-          <CssReset />
-          <Base />
-          <ComponentStyles />
-          {state.rootLayoutMissingTags?.length ? (
-            <RootLayoutMissingTagsError
-              missingTags={state.rootLayoutMissingTags}
-            />
-          ) : hasBuildError ? (
-            <BuildError
-              message={state.buildError!}
-              versionInfo={state.versionInfo}
-            />
-          ) : (
-            <>
-              {hasRuntimeErrors ? (
-                <Errors
-                  isAppDir={true}
-                  initialDisplayState={
-                    isReactError ? 'fullscreen' : 'minimized'
-                  }
-                  errors={state.errors}
-                  versionInfo={state.versionInfo}
-                  hasStaticIndicator={hasStaticIndicator}
-                  debugInfo={debugInfo}
-                />
-              ) : null}
-
-              {hasStaticIndicator && (
-                <StaticIndicator dispatcher={dispatcher} />
-              )}
-            </>
-          )}
-        </ShadowPortal>
-      </>
-    )
-  }
+  return <Overlay {...props} />
 }
